code/registration/requirements.txt
Flask
boto3
grpcio
grpcio-tools
Flask-Session
pycryptodome